feat: add act_now/high_priority to vulnerability_summary and Summary proto
- Amend migration 0030: use INT (not SMALLINT) for cve.priority,
  act_now and high_priority on vulnerability_summary
- RecalculateVulnerabilitySummary: project c.priority in
  resolved_vulnerabilities CTE; add act_now/high_priority COUNT FILTER
  in severity_counts; thread through INSERT/ON CONFLICT
- vulnerbility_summary.sql: add act_now/high_priority to
  vulnerability_data CTE (ListVulnerabilitySummaries) and SUM columns
  in GetVulnerabilitySummary
- Proto Summary message: add act_now=11, high_priority=12
- summary.go: populate ActNow/HighPriority in all three Summary
  assembly sites
- vulnerabilities.go: widen toCve/safePriority to int32; drop
  safeInt16 helper (no longer needed)

+++ internal/database/sql/cve_summary.sql.go
@@ -151,7 +151,7 @@ type ListCveSummariesRow struct {
 	EpssPercentile     *float64
 	HasKevEntry        bool
 	KnownRansomwareUse bool
-	Priority           int16
+	Priority           int32
 	AffectedWorkloads  int32
 	TotalCount         int32
 }
